get d'un element funciona parcialment
A falta d'obtenir totes les dades correctament, es poden mostrar les que hi ha disponibles al fragment de viewElement
