+++ app/src/main/java/edu/upc/pes/wallachange/APILayer/AdapterAPIRequest.java
@@ -40,9 +40,7 @@ public class AdapterAPIRequest   {
 
 
     public void GETRequestAPI(String url, Response.Listener responseListener, Response.ErrorListener errorListener, final Map<String,String> headers){
-
         String  REQUEST_TAG = "com.androidtutorialpoint.volleyJsonObjectRequest";
-
         JsonObjectRequest jsonObjectReq = new JsonObjectRequest(Request.Method.GET, url, null, responseListener, errorListener) {
             @Override
             public Map<String, String> getHeaders() throws AuthFailureError {
@@ -53,6 +51,8 @@ public class AdapterAPIRequest   {
         };
         AppSingleton.getInstance(getApplicationContext()).addToRequestQueue(jsonObjectReq,REQUEST_TAG);
     }
+
+
 
 
     public void GETJsonArrayRequestAPI(String url, Response.Listener responseListener, Response.ErrorListener errorListener,  final JSONObject body, final JSONObject capceleres){
